refactor: Rename `lib/plugins/aws/utils/formatLambdaLogEvent.js`
