+++ lib/plugins/aws/invoke.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const validate = require('./lib/validate');
 const stdin = require('get-stdin');
-const formatLambdaLogEvent = require('./utils/formatLambdaLogEvent');
+const formatLambdaLogEvent = require('./utils/format-lambda-log-event');
 const ServerlessError = require('../../serverless-error');
 const { writeText, style } = require('@serverless/utils/log');
