+++ lib/plugins/aws/logs.js
@@ -6,7 +6,7 @@ const utc = require('dayjs/plugin/utc');
 const wait = require('timers-ext/promise/sleep');
 const { writeText } = require('@serverless/utils/log');
 const validate = require('./lib/validate');
-const formatLambdaLogEvent = require('./utils/formatLambdaLogEvent');
+const formatLambdaLogEvent = require('./utils/format-lambda-log-event');
 const ServerlessError = require('../../serverless-error');
 
 dayjs.extend(utc);
