+++ lib/plugins/aws/utils/format-lambda-log-event.js
